Create Clean up project folder
This script creates a new folder for each composition in the project and moves the composition files to their respective folders. It then creates new folders for each asset type (footage, audio, and other) and moves the respective files to their respective folders. Finally, it collects the project and consolidates the footage, which copies all the project files and footage to a new folder and replaces all the footage items with a single instance of the footage file. This helps to clean up the project folder and organize the files in a more structured and efficient way. Of course, this is just an example and can be modified to suit your specific needs. For example, you can change the asset types that are included or add additional folders or sorting criteria. Additionally, this script assumes that the project has been saved and has a valid file path, so you will need to save the project before running the script.